Move mf6.utilities.hfb to util.hfb and revert change made to hfb mask
