+++ imod/tests/test_mf6/test_utilities/test_hfb.py
@@ -1,7 +1,7 @@
 import pandas as pd
 import pytest
 
-from imod.mf6.utilities.hfb import _prepare_index_names
+from imod.util.hfb import _prepare_index_names
 
 
 def test_pandas_behavior_index_naming_expected():
